finish N-API
