+++ crates/edr_napi/index.d.ts
@@ -63,95 +63,6 @@ export interface CallOverrideResult {
   result: Buffer
   shouldRevert: boolean
 }
-/** Identifier for the Ethereum spec. */
-export const enum SpecId {
-  /** Frontier */
-  Frontier = 0,
-  /** Frontier Thawing */
-  FrontierThawing = 1,
-  /** Homestead */
-  Homestead = 2,
-  /** DAO Fork */
-  DaoFork = 3,
-  /** Tangerine */
-  Tangerine = 4,
-  /** Spurious Dragon */
-  SpuriousDragon = 5,
-  /** Byzantium */
-  Byzantium = 6,
-  /** Constantinople */
-  Constantinople = 7,
-  /** Petersburg */
-  Petersburg = 8,
-  /** Istanbul */
-  Istanbul = 9,
-  /** Muir Glacier */
-  MuirGlacier = 10,
-  /** Berlin */
-  Berlin = 11,
-  /** London */
-  London = 12,
-  /** Arrow Glacier */
-  ArrowGlacier = 13,
-  /** Gray Glacier */
-  GrayGlacier = 14,
-  /** Merge */
-  Merge = 15,
-  /** Shanghai */
-  Shanghai = 16,
-  /** Cancun */
-  Cancun = 17,
-  /** Latest */
-  Latest = 18
-}
-export interface DebugTraceResult {
-  pass: boolean
-  gasUsed: bigint
-  output?: Buffer
-  structLogs: Array<DebugTraceLogItem>
-}
-export interface DebugTraceLogItem {
-  /** Program Counter */
-  pc: bigint
-  op: number
-  /** Gas left before executing this operation as hex number. */
-  gas: string
-  /** Gas cost of this operation as hex number. */
-  gasCost: string
-  /** Array of all values (hex numbers) on the stack */
-  stack?: Array<string>
-  /** Depth of the call stack */
-  depth: bigint
-  /** Size of memory array */
-  memSize: bigint
-  /** Name of the operation */
-  opName: string
-  /** Description of an error as a hex string. */
-  error?: string
-  /** Array of all allocated values as hex strings. */
-  memory?: Array<string>
-  /** Map of all stored values with keys and values encoded as hex strings. */
-  storage?: Record<string, string>
-}
-/** Ethereum execution log. */
-export interface ExecutionLog {
-  address: Buffer
-  topics: Array<Buffer>
-  data: Buffer
-}
-export interface ContractAndFunctionName {
-  /** The contract name. */
-  contractName: string
-  /** The function name. Only present for calls. */
-  functionName?: string
-}
-export interface LoggerConfig {
-  /** Whether to enable the logger. */
-  enable: boolean
-  decodeConsoleLogInputsCallback: (inputs: Buffer[]) => string[]
-  getContractAndFunctionNameCallback: (code: Buffer, calldata?: Buffer) => ContractAndFunctionName
-  printLineCallback: (message: string, replace: boolean) => void
-}
 /** Configuration for a chain */
 export interface ChainConfig {
   /** The chain ID */
@@ -180,7 +91,7 @@ export interface HardforkActivation {
   /** The block number at which the hardfork is activated */
   blockNumber: bigint
   /** The activated hardfork */
-  specId: SpecId
+  specId: string
 }
 /**The type of ordering to use when selecting blocks to mine. */
 export const enum MineOrdering {
@@ -233,7 +144,7 @@ export interface ProviderConfig {
   /** The genesis accounts of the blockchain */
   genesisAccounts: Array<GenesisAccount>
   /** The hardfork of the blockchain */
-  hardfork: SpecId
+  hardfork: string
   /**
    * The initial base fee per gas of the blockchain. Required for EIP-1559
    * transactions and later
@@ -254,6 +165,97 @@ export interface ProviderConfig {
   mining: MiningConfig
   /** The network ID of the blockchain */
   networkId: bigint
+}
+export interface DebugTraceResult {
+  pass: boolean
+  gasUsed: bigint
+  output?: Buffer
+  structLogs: Array<DebugTraceLogItem>
+}
+export interface DebugTraceLogItem {
+  /** Program Counter */
+  pc: bigint
+  op: number
+  /** Gas left before executing this operation as hex number. */
+  gas: string
+  /** Gas cost of this operation as hex number. */
+  gasCost: string
+  /** Array of all values (hex numbers) on the stack */
+  stack?: Array<string>
+  /** Depth of the call stack */
+  depth: bigint
+  /** Size of memory array */
+  memSize: bigint
+  /** Name of the operation */
+  opName: string
+  /** Description of an error as a hex string. */
+  error?: string
+  /** Array of all allocated values as hex strings. */
+  memory?: Array<string>
+  /** Map of all stored values with keys and values encoded as hex strings. */
+  storage?: Record<string, string>
+}
+export const L1_CHAIN_TYPE: string
+export function l1ProviderFactory(): ProviderFactory
+/** Identifier for the Ethereum spec. */
+export const enum SpecId {
+  /** Frontier */
+  Frontier = 0,
+  /** Frontier Thawing */
+  FrontierThawing = 1,
+  /** Homestead */
+  Homestead = 2,
+  /** DAO Fork */
+  DaoFork = 3,
+  /** Tangerine */
+  Tangerine = 4,
+  /** Spurious Dragon */
+  SpuriousDragon = 5,
+  /** Byzantium */
+  Byzantium = 6,
+  /** Constantinople */
+  Constantinople = 7,
+  /** Petersburg */
+  Petersburg = 8,
+  /** Istanbul */
+  Istanbul = 9,
+  /** Muir Glacier */
+  MuirGlacier = 10,
+  /** Berlin */
+  Berlin = 11,
+  /** London */
+  London = 12,
+  /** Arrow Glacier */
+  ArrowGlacier = 13,
+  /** Gray Glacier */
+  GrayGlacier = 14,
+  /** Merge */
+  Merge = 15,
+  /** Shanghai */
+  Shanghai = 16,
+  /** Cancun */
+  Cancun = 17,
+  /** Latest */
+  Latest = 18
+}
+/** Ethereum execution log. */
+export interface ExecutionLog {
+  address: Buffer
+  topics: Array<Buffer>
+  data: Buffer
+}
+export interface ContractAndFunctionName {
+  /** The contract name. */
+  contractName: string
+  /** The function name. Only present for calls. */
+  functionName?: string
+}
+export interface LoggerConfig {
+  /** Whether to enable the logger. */
+  enable: boolean
+  decodeConsoleLogInputsCallback: (inputs: Buffer[]) => string[]
+  getContractAndFunctionNameCallback: (code: Buffer, calldata?: Buffer) => ContractAndFunctionName
+  printLineCallback: (message: string, replace: boolean) => void
 }
 /** The possible reasons for successful termination of the EVM. */
 export const enum SuccessReason {
@@ -341,6 +343,11 @@ export interface ExecutionResult {
   /** Optional contract address if the transaction created a new contract. */
   contractAddress?: Buffer
 }
+/** Configuration for subscriptions. */
+export interface SubscriptionConfig {
+  /** Callback to be called when a new event is received. */
+  subscriptionCallback: (event: SubscriptionEvent) => void
+}
 export interface SubscriptionEvent {
   filterId: bigint
   result: any
@@ -401,26 +408,31 @@ export interface Withdrawal {
 export class EdrContext {
   /**Creates a new [`EdrContext`] instance. Should only be called once! */
   constructor()
+  /**Constructs a new provider with the provided configuration. */
+  createProvider(chainType: string, providerConfig: ProviderConfig, loggerConfig: LoggerConfig, subscriptionConfig: SubscriptionConfig): Promise<Provider>
+  /**Registers a new provider factory for the provided chain type. */
+  registerProviderFactory(chainType: string, factory: ProviderFactory): Promise<void>
 }
+export class ProviderFactory { }
 /** A JSON-RPC provider for Ethereum. */
 export class Provider {
-  /**Constructs a new provider with the provided configuration. */
-  static withConfig(context: EdrContext, config: ProviderConfig, loggerConfig: LoggerConfig, subscriberCallback: (event: SubscriptionEvent) => void): Promise<Provider>
   /**Handles a JSON-RPC request and returns a JSON-RPC response. */
-  handleRequest(jsonRequest: string): Promise<Response>
-  setCallOverrideCallback(callOverrideCallback: (contract_address: Buffer, data: Buffer) => Promise<CallOverrideResult | undefined>): void
+  handleRequest(request: any): Promise<Response>
+  setCallOverrideCallback(callOverrideCallback: (contract_address: Buffer, data: Buffer) => Promise<CallOverrideResult | undefined>): Promise<void>
   /**
    * Set to `true` to make the traces returned with `eth_call`,
    * `eth_estimateGas`, `eth_sendRawTransaction`, `eth_sendTransaction`,
    * `evm_mine`, `hardhat_mine` include the full stack and memory. Set to
    * `false` to disable this.
    */
-  setVerboseTracing(verboseTracing: boolean): void
+  setVerboseTracing(verboseTracing: boolean): Promise<void>
 }
 export class Response {
-  /** Returns the response data as a JSON string or a JSON object. */
+  /**Returns the response data as a JSON string or a JSON object. */
   get data(): string | any
+  /**Returns the Solidity trace of the transaction that failed to execute, if any. */
   get solidityTrace(): RawTrace | null
+  /**Returns the raw traces of executed contracts. This maybe contain zero or more traces. */
   get traces(): Array<RawTrace>
 }
 export class RawTrace {
