+++ crates/edr_napi/src/cast.rs
