+++ crates/edr_napi/src/log.rs
