+++ crates/edr_napi/src/logger.rs
