+++ crates/edr_napi/src/provider/factory.rs
@@ -1,7 +1,8 @@
 use std::sync::Arc;
 
-use edr_napi_core::provider::SyncProviderFactory;
 use napi_derive::napi;
+
+use crate::{logger::LoggerConfig, provider, subscription::SubscriptionConfig};
 
 #[napi]
 pub struct ProviderFactory {
@@ -19,4 +20,16 @@ impl From<Arc<dyn SyncProviderFactory>> for ProviderFactory {
     fn from(inner: Arc<dyn SyncProviderFactory>) -> Self {
         Self { inner }
     }
+}
+
+/// Trait for creating a new provider using the builder pattern.
+pub trait SyncProviderFactory: Send + Sync {
+    /// Creates a `ProviderBuilder` that.
+    fn create_provider_builder(
+        &self,
+        env: &napi::Env,
+        provider_config: provider::Config,
+        logger_config: LoggerConfig,
+        subscription_config: SubscriptionConfig,
+    ) -> napi::Result<Box<dyn provider::Builder>>;
 }
