--- crates/edr_napi/src/provider/optimism.rs
@@ -1,1 +1,0 @@
-
