+++ crates/edr_napi/test/provider.ts
@@ -4,9 +4,9 @@ import chaiAsPromised from "chai-as-promised";
 import {
   ContractAndFunctionName,
   EdrContext,
+  L1_CHAIN_TYPE,
+  l1ProviderFactory,
   MineOrdering,
-  Provider,
-  SpecId,
   SubscriptionEvent,
 } from "..";
 import { collectMessages, collectSteps, ALCHEMY_URL } from "./helpers";
@@ -15,6 +15,11 @@ chai.use(chaiAsPromised);
 
 describe("Provider", () => {
   const context = new EdrContext();
+
+  before(async () => {
+    await context.registerProviderFactory(L1_CHAIN_TYPE, l1ProviderFactory());
+  });
+
   const providerConfig = {
     allowBlocksWithSameTimestamp: false,
     allowUnlimitedContractSize: true,
@@ -32,7 +37,7 @@ describe("Provider", () => {
         balance: 1000n * 10n ** 18n,
       },
     ],
-    hardfork: SpecId.Latest,
+    hardfork: "Latest",
     initialBlobGas: {
       gasUsed: 0n,
       excessGas: 0n,
@@ -68,11 +73,13 @@ describe("Provider", () => {
   };
 
   it("initialize local", async function () {
-    const provider = Provider.withConfig(
-      context,
+    const provider = context.createProvider(
+      L1_CHAIN_TYPE,
       providerConfig,
       loggerConfig,
-      (_event: SubscriptionEvent) => {},
+      {
+        subscriptionCallback: (_event: SubscriptionEvent) => {},
+      },
     );
 
     await assert.isFulfilled(provider);
@@ -83,8 +90,8 @@ describe("Provider", () => {
       this.skip();
     }
 
-    const provider = Provider.withConfig(
-      context,
+    const provider = context.createProvider(
+      L1_CHAIN_TYPE,
       {
         fork: {
           jsonRpcUrl: ALCHEMY_URL,
@@ -92,7 +99,9 @@ describe("Provider", () => {
         ...providerConfig,
       },
       loggerConfig,
-      (_event: SubscriptionEvent) => {},
+      {
+        subscriptionCallback: (_event: SubscriptionEvent) => {},
+      },
     );
 
     await assert.isFulfilled(provider);
@@ -100,30 +109,31 @@ describe("Provider", () => {
 
   describe("verbose mode", function () {
     it("should only include the top of the stack by default", async function () {
-      const provider = await Provider.withConfig(
-        context,
+      const provider = await context.createProvider(
+        L1_CHAIN_TYPE,
         providerConfig,
         loggerConfig,
-        (_event: SubscriptionEvent) => {},
+
+        {
+          subscriptionCallback: (_event: SubscriptionEvent) => {},
+        },
       );
 
-      const responseObject = await provider.handleRequest(
-        JSON.stringify({
-          id: 1,
-          jsonrpc: "2.0",
-          method: "eth_sendTransaction",
-          params: [
-            {
-              from: "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266",
-              // PUSH1 1
-              // PUSH1 2
-              // PUSH1 3
-              // STOP
-              data: "0x60016002600300",
-            },
-          ],
-        }),
-      );
+      const responseObject = await provider.handleRequest({
+        id: 1,
+        jsonrpc: "2.0",
+        method: "eth_sendTransaction",
+        params: [
+          {
+            from: "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266",
+            // PUSH1 1
+            // PUSH1 2
+            // PUSH1 3
+            // STOP
+            data: "0x60016002600300",
+          },
+        ],
+      });
 
       const rawTraces = responseObject.traces;
       assert.lengthOf(rawTraces, 1);
@@ -140,32 +150,34 @@ describe("Provider", () => {
     });
 
     it("should only include the whole stack if verbose mode is enabled", async function () {
-      const provider = await Provider.withConfig(
-        context,
+      const provider = await context.createProvider(
+        L1_CHAIN_TYPE,
         providerConfig,
         loggerConfig,
-        (_event: SubscriptionEvent) => {},
+        {
+          subscriptionCallback: (_event: SubscriptionEvent) => {},
+        },
       );
 
       provider.setVerboseTracing(true);
 
-      const responseObject = await provider.handleRequest(
-        JSON.stringify({
-          id: 1,
-          jsonrpc: "2.0",
-          method: "eth_sendTransaction",
-          params: [
-            {
-              from: "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266",
-              // PUSH1 1
-              // PUSH1 2
-              // PUSH1 3
-              // STOP
-              data: "0x60016002600300",
-            },
-          ],
-        }),
-      );
+      const responseObject = await provider.handleRequest({
+        id: 1,
+        jsonrpc: "2.0",
+        method: "eth_sendTransaction",
+        params: [
+          {
+            from: "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266",
+            // PUSH1 1
+            // PUSH1 2
+            // PUSH1 3
+            // STOP
+            data: "0x60016002600300",
+          },
+        ],
+      });
+
+      console.log(responseObject);
 
       const rawTraces = responseObject.traces;
       assert.lengthOf(rawTraces, 1);
@@ -185,31 +197,31 @@ describe("Provider", () => {
     });
 
     it("should not include memory by default", async function () {
-      const provider = await Provider.withConfig(
-        context,
+      const provider = await context.createProvider(
+        L1_CHAIN_TYPE,
         providerConfig,
         loggerConfig,
-        (_event: SubscriptionEvent) => {},
+        {
+          subscriptionCallback: (_event: SubscriptionEvent) => {},
+        },
       );
 
-      const responseObject = await provider.handleRequest(
-        JSON.stringify({
-          id: 1,
-          jsonrpc: "2.0",
-          method: "eth_sendTransaction",
-          params: [
-            {
-              from: "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266",
-              // store 0x000...001 as the first memory word
-              // PUSH1 1
-              // PUSH0
-              // MSTORE
-              // STOP
-              data: "0x60015f5200",
-            },
-          ],
-        }),
-      );
+      const responseObject = await provider.handleRequest({
+        id: 1,
+        jsonrpc: "2.0",
+        method: "eth_sendTransaction",
+        params: [
+          {
+            from: "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266",
+            // store 0x000...001 as the first memory word
+            // PUSH1 1
+            // PUSH0
+            // MSTORE
+            // STOP
+            data: "0x60015f5200",
+          },
+        ],
+      });
 
       const rawTraces = responseObject.traces;
       assert.lengthOf(rawTraces, 1);
@@ -224,33 +236,33 @@ describe("Provider", () => {
     });
 
     it("should include memory if verbose mode is enabled", async function () {
-      const provider = await Provider.withConfig(
-        context,
+      const provider = await context.createProvider(
+        L1_CHAIN_TYPE,
         providerConfig,
         loggerConfig,
-        (_event: SubscriptionEvent) => {},
+        {
+          subscriptionCallback: (_event: SubscriptionEvent) => {},
+        },
       );
 
       provider.setVerboseTracing(true);
 
-      const responseObject = await provider.handleRequest(
-        JSON.stringify({
-          id: 1,
-          jsonrpc: "2.0",
-          method: "eth_sendTransaction",
-          params: [
-            {
-              from: "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266",
-              // store 0x000...001 as the first memory word
-              // PUSH1 1
-              // PUSH0
-              // MSTORE
-              // STOP
-              data: "0x60015f5200",
-            },
-          ],
-        }),
-      );
+      const responseObject = await provider.handleRequest({
+        id: 1,
+        jsonrpc: "2.0",
+        method: "eth_sendTransaction",
+        params: [
+          {
+            from: "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266",
+            // store 0x000...001 as the first memory word
+            // PUSH1 1
+            // PUSH0
+            // MSTORE
+            // STOP
+            data: "0x60015f5200",
+          },
+        ],
+      });
 
       const rawTraces = responseObject.traces;
       assert.lengthOf(rawTraces, 1);
@@ -270,29 +282,30 @@ describe("Provider", () => {
     });
 
     it("should include isStaticCall flag in tracing messages", async function () {
-      const provider = await Provider.withConfig(
-        context,
+      const provider = await context.createProvider(
+        L1_CHAIN_TYPE,
         providerConfig,
         loggerConfig,
-        (_event: SubscriptionEvent) => {},
+
+        {
+          subscriptionCallback: (_event: SubscriptionEvent) => {},
+        },
       );
 
-      const responseObject = await provider.handleRequest(
-        JSON.stringify({
-          id: 1,
-          jsonrpc: "2.0",
-          method: "eth_sendTransaction",
-          params: [
-            {
-              from: "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266",
-              // make a static call to the zero address
-              // yul: staticcall(gas(), 0, 0, 0, 0, 0)
-              data: "0x6000808080805afa00",
-              gas: "0x" + 1_000_000n.toString(16),
-            },
-          ],
-        }),
-      );
+      const responseObject = await provider.handleRequest({
+        id: 1,
+        jsonrpc: "2.0",
+        method: "eth_sendTransaction",
+        params: [
+          {
+            from: "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266",
+            // make a static call to the zero address
+            // yul: staticcall(gas(), 0, 0, 0, 0, 0)
+            data: "0x6000808080805afa00",
+            gas: "0x" + 1_000_000n.toString(16),
+          },
+        ],
+      });
 
       const rawTraces = responseObject.traces;
       assert.lengthOf(rawTraces, 1);
@@ -309,33 +322,34 @@ describe("Provider", () => {
     });
 
     it("should have tracing information when debug_traceTransaction is used", async function () {
-      const provider = await Provider.withConfig(
-        context,
+      const provider = await context.createProvider(
+        L1_CHAIN_TYPE,
         providerConfig,
         loggerConfig,
-        (_event: SubscriptionEvent) => {},
+
+        {
+          subscriptionCallback: (_event: SubscriptionEvent) => {},
+        },
       );
 
-      const sendTxResponse = await provider.handleRequest(
-        JSON.stringify({
-          id: 1,
-          jsonrpc: "2.0",
-          method: "eth_sendTransaction",
-          params: [
-            {
-              from: "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266",
-              // PUSH1 0x42
-              // PUSH0
-              // MSTORE
-              // PUSH1 0x20
-              // PUSH0
-              // RETURN
-              data: "0x60425f5260205ff3",
-              gas: "0x" + 1_000_000n.toString(16),
-            },
-          ],
-        }),
-      );
+      const sendTxResponse = await provider.handleRequest({
+        id: 1,
+        jsonrpc: "2.0",
+        method: "eth_sendTransaction",
+        params: [
+          {
+            from: "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266",
+            // PUSH1 0x42
+            // PUSH0
+            // MSTORE
+            // PUSH1 0x20
+            // PUSH0
+            // RETURN
+            data: "0x60425f5260205ff3",
+            gas: "0x" + 1_000_000n.toString(16),
+          },
+        ],
+      });
 
       let responseData;
 
@@ -347,47 +361,46 @@ describe("Provider", () => {
 
       const txHash = responseData.result;
 
-      const traceTransactionResponse = await provider.handleRequest(
-        JSON.stringify({
-          id: 1,
-          jsonrpc: "2.0",
-          method: "debug_traceTransaction",
-          params: [txHash],
-        }),
-      );
+      const traceTransactionResponse = await provider.handleRequest({
+        id: 1,
+        jsonrpc: "2.0",
+        method: "debug_traceTransaction",
+        params: [txHash],
+      });
 
       const rawTraces = traceTransactionResponse.traces;
       assert.lengthOf(rawTraces, 1);
     });
 
     it("should have tracing information when debug_traceCall is used", async function () {
-      const provider = await Provider.withConfig(
-        context,
+      const provider = await context.createProvider(
+        L1_CHAIN_TYPE,
         providerConfig,
         loggerConfig,
-        (_event: SubscriptionEvent) => {},
+
+        {
+          subscriptionCallback: (_event: SubscriptionEvent) => {},
+        },
       );
 
-      const traceCallResponse = await provider.handleRequest(
-        JSON.stringify({
-          id: 1,
-          jsonrpc: "2.0",
-          method: "debug_traceCall",
-          params: [
-            {
-              from: "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266",
-              // PUSH1 0x42
-              // PUSH0
-              // MSTORE
-              // PUSH1 0x20
-              // PUSH0
-              // RETURN
-              data: "0x60425f5260205ff3",
-              gas: "0x" + 1_000_000n.toString(16),
-            },
-          ],
-        }),
-      );
+      const traceCallResponse = await provider.handleRequest({
+        id: 1,
+        jsonrpc: "2.0",
+        method: "debug_traceCall",
+        params: [
+          {
+            from: "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266",
+            // PUSH1 0x42
+            // PUSH0
+            // MSTORE
+            // PUSH1 0x20
+            // PUSH0
+            // RETURN
+            data: "0x60425f5260205ff3",
+            gas: "0x" + 1_000_000n.toString(16),
+          },
+        ],
+      });
 
       const rawTraces = traceCallResponse.traces;
       assert.lengthOf(rawTraces, 1);
