--- crates/edr_napi_core/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "edr_napi_core"
-version = "0.3.5"
-edition = "2021"
-
-[dependencies]
-ansi_term = { version = "0.12.1", default-features = false }
-derive-where = { version = "1.2.7", default-features = false }
-edr_defaults = { path = "../edr_defaults" }
-edr_eth = { path = "../edr_eth" }
-edr_evm = { path = "../edr_evm" }
-edr_provider = { path = "../edr_provider" }
-edr_rpc_client = { path = "../edr_rpc_client" }
-itertools = { version = "0.12.0", default-features = false }
-# when napi is pinned, be sure to pin napi-derive to the same version
-# The `async` feature ensures that a tokio runtime is available
-napi = { version = "2.16.0", default-features = false, features = ["async", "error_anyhow", "napi8", "serde-json"] }
-napi-derive = "2.16.0"
-serde = { version = "1.0.189", features = ["derive"] }
-serde_json = { version = "1.0.89" }
-thiserror = { version = "1.0.37", default-features = false }
-
-[lints]
-workspace = true
