--- crates/edr_napi_core/src/lib.rs
@@ -1,18 +1,0 @@
-#![warn(missing_docs)]
-
-/// Types for overriding a call.
-pub mod call_override;
-/// Types for casting N-API types to Rust types.
-pub mod cast;
-/// Types for EVM execution logs.
-pub mod log;
-/// Types for an RPC request logger.
-pub mod logger;
-/// Types for Ethereum RPC providers.
-pub mod provider;
-/// Types for EVM execution results.
-pub mod result;
-/// Types for subscribing to events.
-pub mod subscription;
-/// Types for EVM traces.
-pub mod trace;
