+++ crates/edr_provider/src/test_utils.rs
@@ -1,11 +1,13 @@
 use std::{num::NonZeroU64, time::SystemTime};
 
 use edr_eth::{
-    block::BlobGas, chain_spec::L1ChainSpec, signature::secret_key_from_str, trie::KECCAK_NULL_RLP,
+    block::BlobGas, chain_spec::L1ChainSpec, result::InvalidTransaction,
+    signature::secret_key_from_str, transaction::TransactionValidation, trie::KECCAK_NULL_RLP,
     Address, Bytes, HashMap, B256, U160, U256,
 };
 use edr_evm::Block;
 use edr_rpc_eth::TransactionRequest;
+use time::TimeSinceEpoch;
 
 use super::*;
 use crate::{config::MiningConfig, requests::hardhat::rpc_types::ForkConfig};
@@ -98,7 +100,7 @@ pub fn pending_base_fee<
 /// Deploys a contract with the provided code. Returns the address of the
 /// contract.
 pub fn deploy_contract<TimerT>(
-    provider: &Sequential<L1ChainSpec, TimerT>,
+    provider: &Provider<L1ChainSpec, TimerT>,
     caller: Address,
     code: Bytes,
 ) -> anyhow::Result<Address>
