+++ crates/edr_provider/tests/issues/issue_533.rs
@@ -3,7 +3,7 @@ use std::str::FromStr as _;
 use edr_eth::{chain_spec::L1ChainSpec, B256};
 use edr_provider::{
     hardhat_rpc_types::ForkConfig, test_utils::create_test_config_with_fork, time::CurrentTime,
-    MethodInvocation, NoopLogger, Sequential, ProviderRequest,
+    MethodInvocation, NoopLogger, Provider, ProviderRequest,
 };
 use edr_test_utils::env::get_alchemy_url;
 use tokio::runtime;
@@ -23,7 +23,7 @@ async fn issue_533() -> anyhow::Result<()> {
     // The default chain id set by Hardhat
     config.chain_id = 31337;
 
-    let provider = Sequential::new(
+    let provider = Provider::new(
         runtime::Handle::current(),
         logger,
         subscriber,
